Rename unknown_078050 to rcp.
