+++ src/game_ui.c
@@ -16,7 +16,7 @@
 #include "game_text.h"
 #include "object_models.h"
 #include "tracks.h"
-#include "unknown_078050.h"
+#include "rcp.h"
 #include "audiosfx.h"
 #include "printf.h"
 #include "unknown_008C40.h"
@@ -288,7 +288,7 @@ s32 D_80127190;
 s32 D_80127194;
 s32 D_80127198[6];
 
-// Not sure why this ended up here, and not in unknown_078050.c along with the rest of the task data.
+// Not sure why this ended up here, and not in rcp.c along with the rest of the task data.
 u8 gGfxTaskYieldData[OS_YIELD_DATA_SIZE];
 
 /******************************/
