+++ src/menu.c
@@ -22,12 +22,11 @@
 #include "thread30_track_loading.h"
 #include "objects.h"
 #include "game.h"
-#include "unknown_078050.h"
+#include "rcp.h"
 #include "save_data.h"
 #include "object_functions.h"
 #include "audiosfx.h"
 #include "racer.h"
-#include "unknown_078050.h"
 #include "tracks.h"
 #include "lib/src/mips1/al/alSynStartVoiceParams.h"
 #include "controller.h"
